Updating league_builder.py
This is when I realized that since the csv unpacking stuff distributed the CSV's heading into every ordereddict in my list, that there was no need for my heading and rosters variables.  Also, the only reason i'd gotten "so far" was that i'd accidentally called for row in rows rather than row in roster.  rows[0] is already my first player, rather than a list of keys.